requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile --output-file=requirements.txt --resolver=backtracking requirements.in
#
black==23.3.0
    # via -r requirements.in
certifi==2022.12.7
    # via requests
charset-normalizer==3.1.0
    # via requests
click==8.1.3
    # via black
coverage==7.2.3
    # via -r requirements.in
exceptiongroup==1.1.1
    # via pytest
idna==3.4
    # via requests
iniconfig==2.0.0
    # via pytest
mypy-extensions==1.0.0
    # via black
numpy==1.24.2
    # via
    #   pandas
    #   pyarrow
packaging==23.0
    # via
    #   black
    #   pytest
pandas==2.0.0
    # via -r requirements.in
pathspec==0.11.1
    # via black
platformdirs==3.2.0
    # via black
pluggy==1.0.0
    # via pytest
pyarrow==11.0.0
    # via -r requirements.in
pytest==7.3.0
    # via -r requirements.in
python-dateutil==2.8.2
    # via pandas
pytz==2023.3
    # via pandas
requests==2.28.2
    # via -r requirements.in
six==1.16.0
    # via python-dateutil
tomli==2.0.1
    # via
    #   black
    #   pytest
typing-extensions==4.5.0
    # via black
tzdata==2023.3
    # via pandas
urllib3==1.26.15
    # via requests
